My First Android Application
The application uses Google Maps to locate and show other users. The
user has the option to post a message to be shown on the map, but also a
message that is shown in a Live feed. The application transfers data
using a server and stores the data in JSON-strings.